+++ CHANGELOG.md
@@ -10,6 +10,7 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ### Added
 
 * Added support for the setting of endpointConfiguration to be deferred until after agent initialization. Disabling the endpoint is also supported with caching of pending spans.
+* XCFramework build system for binary distribution of the SDK.
 
 ### Changed